Rename and move new journal entry form page inside of dashboard folder
