+++ src/pages/DashboardPage/DashboardNewJournalEntryFormPage/DashboardNewJournalEntryFormPage.css
